Add version test of yaml module when testing
The JSON::Validator seems to require at least version 0.67 of YAML::XS
to work. If an older version is installed the the test fails. I've
changed the pre-condition to check for a specific version of YAML::XS,
and if it's not installed the test is skipped, just as if the modules
wasn't installed at all.

This should most/all of the CPAN Testers failures.

This is part of the CPAN Pull Request Challenge, where I got